Fix #16 - Remove the fs.exists check for handling USR2 signals
When running multiple processes that are writing to the same log
(not clustered), there is a race condition where the first process
to handle the signal will reopen the file handle (creating the
file), and all other processes will ignore the request because the
file already exists.  This fixes that issue.